[^]-Fix negative number checking.

+++ Sources/_Common/Stack/stack_is_sort.c
@@ -6,7 +6,7 @@
 /*   By: fgalaup <fgalaup@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/07 11:58:12 by fgalaup           #+#    #+#             */
-/*   Updated: 2021/05/10 13:36:48 by fgalaup          ###   ########lyon.fr   */
+/*   Updated: 2021/05/11 14:26:54 by fgalaup          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ t_bool	stack_is_sort(t_bidirectional_list *stack)
 	while (it != NULL)
 	{
 		if (it->next != NULL)
-			if (it->content > it->next->content)
+			if ((long long)it->content > (long long)it->next->content)
 				return (FT_FALSE);
 		it = it->next;
 	}
